update filter items
